# digital-gateway-service/src/main/resources/openapi/customer-api.yaml
{ "openapi": "3.1.0","info": { "title": "Customer API","description": "Customer Management Service API","version": "v1" },"servers": [ { "url": "http://localhost:8081","description": "Generated server url" } ],"tags": [ { "name": "Customer","description": "Customer management APIs" } ],"paths": { "/customer": { "post": { "tags": [ "Customer" ],"summary": "Creates a Customer","description": "This operation creates a Customer entity","operationId": "save","requestBody": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CustomerRequestDto" } } },"required": true },"responses": { "201": { "description": "Customer created","content": { "*/*": { "schema": { "$ref": "#/components/schemas/CustomerResponseDto" } } } },"400": { "description": "Invalid request","content": { "*/*": { "schema": { "$ref": "#/components/schemas/CustomerResponseDto" } } } } } } },"/customer/{id}": { "get": { "tags": [ "Customer" ],"summary": "Retrieves a Customer by ID","description": "This operation retrieves a Customer entity. Attribute selection enabled for all first level attributes.","operationId": "get","parameters": [ { "name": "id","in": "path","description": "Customer ID","required": true,"schema": { "type": "string" },"example": 1 } ],"responses": { "200": { "description": "OK","content": { "*/*": { "schema": { "$ref": "#/components/schemas/CustomerResponseDto" } } } } } },"delete": { "tags": [ "Customer" ],"summary": "Deletes a Customer","description": "This operation deletes a Customer entity.","operationId": "delete","parameters": [ { "name": "id","in": "path","required": true,"schema": { "type": "string" } } ],"responses": { "200": { "description": "OK" } } },"patch": { "tags": [ "Customer" ],"summary": "Updates partially a Customer","description": "This operation updates partially a Customer entity.","operationId": "update","parameters": [ { "name": "id","in": "path","description": "Customer ID","required": true,"schema": { "type": "string" },"example": 1 } ],"requestBody": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CustomerRequestDto" } } },"required": true },"responses": { "200": { "description": "OK","content": { "*/*": { "schema": { "$ref": "#/components/schemas/CustomerResponseDto" } } } } } } } },"components": { "schemas": { "CustomerRequestDto": { "type": "object","properties": { "name": { "type": "string","example": "Moon Football Club","minLength": 1 },"description": { "type": "string","example": "Optional description" },"engagedPartyId": { "type": "string","description": "Engaged Party ID reference","example": "53af0a1c-088b-427c-9305-a72f41374415" },"type": { "type": "string","example": "Customer","minLength": 1 } },"required": [ "engagedPartyId","name","type" ] },"CustomerResponseDto": { "type": "object","properties": { "id": { "type": "string","example": "8416647f-7c4b-4487-85e2-30705ebee482" },"name": { "type": "string","example": "Moon Football Club" },"description": { "type": "string","example": "Optional description" },"engagedParty": { "$ref": "#/components/schemas/EngagedPartyDto","example": "Engaged Party" },"type": { "type": "string","example": "Customer" } } },"EngagedPartyDto": { "type": "object","properties": { "id": { "type": "string","example": "53af0a1c-088b-427c-9305-a72f41374415","minLength": 1 },"name": { "type": "string","example": "Happy Travellers" },"type": { "type": "string","example": "PartyRef","minLength": 1 } },"required": [ "id","type" ] } } } }